Clarify scope of include/LICENSE.txt
The existing license text could be misinterpreted as pmdsky-debug owning
everything in the directory, which isn't true. Modify the note to
clarify the scope of the license's application.

As the one named as the original copyright holder, I'm loosening it for
this repository since the portions of work it's using are more naturally
attributed to all the pmdsky-debug contributors, not just me.