Sync some symbols from pokeheartgold
These were found by correlating identical code between Sky and Heart
Gold.

This is the first of a series of results from this correlation work.
Some of the changes included in this commit:
- init code
- math utilities (FX_, VEC_, MTX_)
- graphics utilities (GX_ and friends)

None of the code here is decompiled, so this change only includes
renames and tweaks to the asm.

Note: GX_HBlankIntr is not the same between EoS and Heart Gold, but the
label was added anyway because it is between GX_Init and GX_VBlankIntr
in both games. The latter two functions DO match between games.